# server/api/src/openapi_server/openapi/openapi.yaml
openapi: 3.0.1
info:
  description: Kaldi Customization Server.
  title: Kaldi Customization Server
  version: 1.1.0
servers:
- url: http://localhost:8080/api/v1
- url: https://localhost:8080/api/v1
...
paths:
...
  /project/{project_uuid}/training:
    get:
      operationId: get_trainings_for_project
      parameters:
      - description: Get all trainings for a specific project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Training'
                type: array
          description: all trainings
        403:
          description: Forbidden
        404:
          content: {}
          description: Project not found
      security:
      - oauth:
        - read:projects
      summary: Lists all Trainings of a Project
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
    post:
      operationId: create_training
      parameters:
      - description: Project object that needs to be trained
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
          description: created successfully
        403:
          description: Forbidden
        404:
          content: {}
          description: Project not found
      security:
      - oauth:
        - write:projects
        - read:projects
        - train:projects
      summary: Create a new training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
  /project/{project_uuid}/training/current:
    get:
      operationId: get_current_training_for_project
      parameters:
      - description: Get the current training for a specific project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
          description: The current training
        403:
          description: Forbidden
        404:
          content: {}
          description: Project or Training not found
      security:
      - oauth:
        - read:projects
      summary: Get current training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
  /project/{project_uuid}/training/{training_version}:
    get:
      description: Returns the training object
      operationId: get_training_by_version
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
          description: Got the specified training
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - read:projects
      summary: Find project training results by UUID
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
  /project/{project_uuid}/training/{training_version}/corpus:
    get:
      description: Returns the entire corpus of the specified training
      operationId: get_corpus_of_training
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            text/plain:
              schema:
                description: Corpus of the training
                example: er horte leise schritte hinter sich das bedeutete nichts
                  gutes wer wurde ihm schon folgen spat in der nacht und dazu noch
                  in dieser engen gasse mitten im ubel beleumundeten hafenviertel
                  gerade jetzt wo er das ding seines lebens gedreht hatte und mit
                  der beute verschwinden wollte hatte einer seiner zahllosen kollegen
                  dieselbe idee gehabt ihn beobachtet und abgewartet um ihn nun um
                  die fruchte seiner arbeit zu erleichtern oder gehorten die schritte
                  hinter ihm zu einem der unzahligen gesetzeshuter dieser stadt und
                  die stahlerne acht um seine handgelenke wurde gleich zuschnappen
                  er konnte die aufforderung stehen zu bleiben schon horen gehetzt
                  sah er sich um plotzlich erblickte er den schmalen durchgang blitzartig
                  drehte er sich nach rechts und verschwand zwischen den beiden gebauden
                  beinahe ware er dabei uber den umgesturzten mulleimer gefallen der
                  mitten im weg lag
                type: string
          description: Corpus as plain text
        400:
          content: {}
          description: Training not prepared
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - read:projects
      summary: Get the entire corpus of the specified training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
...
  /project/{project_uuid}/training/{training_version}/model:
    get:
      description: Returns the model of the specified training
      operationId: download_model_for_training
      parameters:
      - description: UUID of project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          type: string
        style: simple
      - description: Version of training
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            application/zip:
              schema:
                format: binary
                type: string
          description: Trained model content
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Page not found
      security:
      - oauth:
        - read:projects
      summary: Returns the model
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
  /project/{project_uuid}/training/{training_version}/prepare:
    put:
      description: Start the preparation process for the specified training
      operationId: prepare_training_by_version
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackObject'
        description: Callback to be executed after the operation ended
        required: false
      responses:
        202:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
          description: Preparation successfully queued
        400:
          content: {}
          description: training already done or pending
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - read:projects
        - train:projects
      summary: Prepare the specified training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
...
  /project/{project_uuid}/training/{training_version}/resource:
    post:
      description: Assign the specified resource to the training
      operationId: assign_resource_to_training
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceReferenceObject'
        description: Resource that needs to be added
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resource'
          description: Resource successfully assigned
        400:
          content: {}
          description: Resource already in training
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project, training or resource not found
      security:
      - oauth:
        - read:projects
        - write:projects
        - read:resources
        - write:resources
      summary: Assign a resource to the training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
  /project/{project_uuid}/training/{training_version}/resource/{resource_uuid}:
    delete:
      description: Removes the assigned resource from the training
      operationId: delete_assigned_resource_from_training
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      - description: UUID of the resource
        explode: false
        in: path
        name: resource_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      responses:
        200:
          content: {}
          description: Resource assignment successfully removed
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project, training or resource not found
        409:
          description: 'Conflict: already in training'
      security:
      - oauth:
        - read:projects
        - write:projects
        - read:resources
        - write:resources
      summary: Remove a resource from the training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
...
  /project/{project_uuid}/training/{training_version}/train:
    put:
      description: Start the training process for the specified training
      operationId: start_training_by_version
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CallbackObject'
        description: Callback to be executed after the operation ended
        required: false
      responses:
        202:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Training'
          description: Training successfully queued
        400:
          content: {}
          description: training already done or pending
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - read:projects
        - train:projects
      summary: Start the specified training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
...
  /project/{project_uuid}/training/{training_version}/vocabulary:
    get:
      description: Returns the entire vocabulary of the specified training
      operationId: get_vocabulary_of_training
      parameters:
      - description: UUID of the project
        explode: false
        in: path
        name: project_uuid
        required: true
        schema:
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
        style: simple
      - description: Training version of the project
        explode: false
        in: path
        name: training_version
        required: true
        schema:
          type: integer
        style: simple
      responses:
        200:
          content:
            text/plain:
              schema:
                description: Vocabulary of the training
                example: wort noch ein bla
                type: string
          description: Vocabulary as line-separated plain text
        400:
          content: {}
          description: Training not prepared
        403:
          content: {}
          description: Forbidden
        404:
          content: {}
          description: Project or training not found
      security:
      - oauth:
        - read:projects
      summary: Get the entire vocabulary of the specified training
      tags:
      - training
      x-openapi-router-controller: openapi_server.controllers.training_controller
...
components:
  schemas:
...
    Training:
      example:
        creation_timestamp: 2000-01-23T04:56:07.000+00:00
        resources:
        - creation_timestamp: 2000-01-23T04:56:07.000+00:00
          name: myFile.pdf
          uuid: 550e8400-e29b-11d4-a716-446655440000
        - creation_timestamp: 2000-01-23T04:56:07.000+00:00
          name: myFile.pdf
          uuid: 550e8400-e29b-11d4-a716-446655440000
        version: 0
      properties:
        version:
          example: 0
          type: integer
        creation_timestamp:
          format: date-time
          type: string
        status:
          $ref: '#/components/schemas/TrainingStatus'
        resources:
          items:
            $ref: '#/components/schemas/Resource'
          type: array
      type: object
...
    ResourceReferenceObject:
      example:
        resource_uuid: 550e8400-e29b-11d4-a716-446655440000
      properties:
        resource_uuid:
          description: UUID of the resource
          example: 550e8400-e29b-11d4-a716-446655440000
          format: uuid
          type: string
      required:
      - resource_uuid
      type: object
...
    CallbackObject:
      example:
        url: url
      properties:
        url:
          description: The URL to be called
          type: string
        method:
          $ref: '#/components/schemas/CallbackMethod'
      required:
      - method
      - url
  securitySchemes:
    oauth:
      flows:
        password:
          scopes:
            write:user: modify the user account
            read:user: read the user account
            write:projects: modify projects in your account
            read:projects: read your projects
            train:projects: train projects in your account
            decode:projects: decode audio data using projects in your account
            write:resources: modify resources in your account
            read:resources: read your resources
            write:audio: modify your audio
            read:audio: read your audio
          tokenUrl: http://localhost:8080/api/oauth/token
      type: oauth2
      x-tokenInfoFunc: openapi_server.controllers.security_controller_.info_from_oauth
      x-scopeValidateFunc: openapi_server.controllers.security_controller_.validate_scope_oauth